+++ sentry-android-core/src/main/java/io/sentry/android/core/AndroidOptionsInitializer.java
@@ -203,8 +203,10 @@ final class AndroidOptionsInitializer {
             new SendFireAndForgetOutboxSender(() -> options.getOutboxPath()),
             hasStartupCrashMarker));
 
-    options.addIntegration(new AnrIntegration(context));
+    // AppLifecycleIntegration has to be installed before AnrIntegration, because AnrIntegration
+    // relies on AppState set by it
     options.addIntegration(new AppLifecycleIntegration());
+    options.addIntegration(new AnrIntegration(context));
 
     // registerActivityLifecycleCallbacks is only available if Context is an AppContext
     if (context instanceof Application) {