+++ sentry/src/main/java/io/sentry/MainEventProcessor.java
@@ -1,5 +1,6 @@
 package io.sentry;
 
+import io.sentry.hints.AbnormalExit;
 import io.sentry.hints.Cached;
 import io.sentry.protocol.DebugImage;
 import io.sentry.protocol.DebugMeta;
@@ -252,8 +253,16 @@ public final class MainEventProcessor implements EventProcessor, Closeable {
         }
       }
 
-      if (options.isAttachThreads()) {
-        event.setThreads(sentryThreadFactory.getCurrentThreads(mechanismThreadIds));
+      // typically Abnormal exits can be tackled by looking at the thread dump (e.g. ANRs), hence
+      // we force attach threads regardless of the config
+      if (options.isAttachThreads() || HintUtils.hasType(hint, AbnormalExit.class)) {
+        final Object sentrySdkHint = HintUtils.getSentrySdkHint(hint);
+        boolean ignoreCurrentThread = false;
+        if (sentrySdkHint instanceof AbnormalExit) {
+          ignoreCurrentThread = ((AbnormalExit) sentrySdkHint).ignoreCurrentThread();
+        }
+        event.setThreads(
+            sentryThreadFactory.getCurrentThreads(mechanismThreadIds, ignoreCurrentThread));
       } else if (options.isAttachStacktrace()
           && (eventExceptions == null || eventExceptions.isEmpty())
           && !isCachedHint(hint)) {